Rename SignedTransactionOrPsbt::Psbt variant to SignedPsbt
Fixes UniFFI name collision in Kotlin where the enum variant name
'Psbt' shadowed the external Psbt type, causing compile errors in
the generated psbt() method.

+++ android/app/src/main/java/org/bitcoinppl/cove_core/device/cove_device.kt
@@ -81,7 +81,7 @@ open class RustBuffer : Structure() {
 
     @Suppress("TooGenericExceptionThrown")
     fun asByteBuffer() =
-        this.data?.getByteBuffer(0, this.len.toLong())?.also {
+        this.data?.getByteBuffer(0, this.len)?.also {
             it.order(ByteOrder.BIG_ENDIAN)
         }
 }
@@ -706,7 +706,7 @@ internal object IntegrityCheckingUniffiLib {
     ): Short
     external fun ffi_cove_device_uniffi_contract_version(
     ): Int
-    
+
         
 }
 
@@ -846,7 +846,7 @@ internal object UniffiLib {
     ): Unit
     external fun ffi_cove_device_rust_future_complete_void(`handle`: Long,uniffi_out_err: UniffiRustCallStatus, 
     ): Unit
-    
+
         
 }
