+++ android/app/src/main/java/org/bitcoinppl/cove_core/ur/cove_ur.kt
@@ -81,7 +81,7 @@ open class RustBuffer : Structure() {
 
     @Suppress("TooGenericExceptionThrown")
     fun asByteBuffer() =
-        this.data?.getByteBuffer(0, this.len.toLong())?.also {
+        this.data?.getByteBuffer(0, this.len)?.also {
             it.order(ByteOrder.BIG_ENDIAN)
         }
 }
@@ -670,7 +670,7 @@ internal object IntegrityCheckingUniffiLib {
     ): Short
     external fun ffi_cove_ur_uniffi_contract_version(
     ): Int
-    
+
         
 }
 
@@ -836,7 +836,7 @@ internal object UniffiLib {
     ): Unit
     external fun ffi_cove_ur_rust_future_complete_void(`handle`: Long,uniffi_out_err: UniffiRustCallStatus, 
     ): Unit
-    
+
         
 }
