+++ android/app/src/main/java/org/bitcoinppl/cove_core/util/cove_util.kt
@@ -80,7 +80,7 @@ open class RustBuffer : Structure() {
 
     @Suppress("TooGenericExceptionThrown")
     fun asByteBuffer() =
-        this.data?.getByteBuffer(0, this.len.toLong())?.also {
+        this.data?.getByteBuffer(0, this.len)?.also {
             it.order(ByteOrder.BIG_ENDIAN)
         }
 }
@@ -645,7 +645,7 @@ internal object IntegrityCheckingUniffiLib {
     ): Short
     external fun ffi_cove_util_uniffi_contract_version(
     ): Int
-    
+
         
 }
 
@@ -768,7 +768,7 @@ internal object UniffiLib {
     ): Unit
     external fun ffi_cove_util_rust_future_complete_void(`handle`: Long,uniffi_out_err: UniffiRustCallStatus, 
     ): Unit
-    
+
         
 }
